requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1==2.7.0
asttokens==2.4.1
bitarray==2.8.5
bitstring==4.1.4
cffi==1.16.0
cryptography==41.0.5
decorator==5.1.1
enum-compat==0.0.3
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.18.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
six==1.16.0
stack-data==0.6.3
traitlets==5.13.0
wcwidth==0.2.12
